Rename src/three -> src/space3d to avoid shadowing the three package
With baseUrl=src, a directory named "three" can shadow bare imports of
the three.js package (one stray index.ts away from a broken build).
