+++ src/space3d/SpaceCanvas.tsx
