+++ src/space3d/StarField.tsx
