+++ src/space3d/starSampling.ts
