+++ src/space3d/textures.ts
